Add standalone version of db_link_password_decrypt
Add a standalone version of db_link_password_decrypt from the
"db_link_password_decrypt.sql" script. This is essentially just a copy
of the existing procedure (minus the CREATE OR REPLACE boilerplate),
embedded into the DECLARE section of an anonymous PL/SQL block. The
primary advantage of which being that it removes the need to create a
persistent database object.

A secondary benefit is that the script will automatically determine the
NO_USERID_VERIFIER_SALT value for the current database, and prompt for
an owner and dblink name. This makes it a bit more user friendly for
the common case, at the expense of being unable to process dblinks from
a different database.